deploys/plotly_app/requirements.txt
ipywidgets==7.7.3
matplotlib
numpy
pandas
plotly
rsconnect-python
shiny
shinywidgets
python-dotenv
